requirements.txt
customtkinter
pillow
tk
pywin32